refactor: share screenshot result typing
