+++ src/client-types.ts
@@ -28,12 +28,8 @@ import type { ScrollInputDirection } from './commands/interaction/runtime/gestur
 import type { LogAction } from './contracts/logs.ts';
 import type { SessionSurface } from './core/session-surface.ts';
 import type { FindLocator } from './utils/finders.ts';
-import type {
-  ScreenshotOverlayRef,
-  SnapshotNode,
-  SnapshotUnchanged,
-  SnapshotVisibility,
-} from './utils/snapshot.ts';
+import type { SnapshotNode, SnapshotUnchanged, SnapshotVisibility } from './utils/snapshot.ts';
+import type { ScreenshotResultData } from './utils/screenshot-result.ts';
 import type {
   MetroPrepareKind,
   PrepareMetroRuntimeResult,
@@ -357,9 +353,8 @@ export type CaptureScreenshotOptions = AgentDeviceRequestOverrides & {
   surface?: SessionSurface;
 };
 
-export type CaptureScreenshotResult = {
+export type CaptureScreenshotResult = ScreenshotResultData & {
   path: string;
-  overlayRefs?: ScreenshotOverlayRef[];
   identifiers: AgentDeviceIdentifiers;
 };
